% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftshare.R
\name{BHJ_IV}
\alias{BHJ_IV}
\title{Perform Borusyak, Hull & Jaravel (2022) transformation to industry-level regression in shift-share IV.}
\usage{
BHJ_IV(outcomes_data, exposure_shares, industry_shocks)
}
\arguments{
\item{outcomes_data}{data.table with columns "y_i", "x_i", and "location"}

\item{exposure_shares}{data.table with columns "industry", "location", and "exposure_share_in"}

\item{industry_shocks}{data.table with columns "industry" and "industry_shock_n"}
}
\description{
Perform Borusyak, Hull & Jaravel (2022) transformation to industry-level regression in shift-share IV.
}
