% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftshare.R
\name{construct_shiftshare_IV}
\alias{construct_shiftshare_IV}
\title{Construct the shift-share IV using industry shocks and exposure shares.}
\usage{
construct_shiftshare_IV(exposure_shares, industry_shocks)
}
\arguments{
\item{exposure_shares}{data.table with columns "industry", "location", and "exposure_share_in"}

\item{industry_shocks}{data.table with columns "industry" and "industry_shock_n"}
}
\description{
Construct the shift-share IV using industry shocks and exposure shares.
}
