% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssa.R
\name{dssa}
\alias{dssa}
\title{Perform Dynamic Shift share Analysis}
\usage{
dssa(df, y.var, region.var, year.var, sector.var, gregion.id, sector.id,
  start.year, end.year)
}
\arguments{
\item{df}{A dataframe}

\item{y.var}{The dependent variable (e.g., employment or growth)}

\item{region.var}{A factor variable that defines regions}

\item{year.var}{A number specifying the year variable}

\item{sector.var}{A factor variable that defines sectors}

\item{gregion.id}{A string that identifies the greater region (e.g., national employment)}

\item{sector.id}{A string that identifies the total sector}

\item{start.year}{A number specifying start year}

\item{end.year}{A number specifying end year}
}
\value{
A dataframe containing the National Growth effect (\code{NSi}),
the Industry Mix effect (\code{IMi}), and the Local Share effect (\code{RSi}) by region and occupation
}
\description{
Perform Dynamic Shift share Analysis
}
\examples{
data('usregions')
my.dssa <- dssa(usregions, y.var = 'TOT_EMP' ,region.var = 'ST', sector.var = 'OCC_TITLE',
  gregion.id = 'US', year.var = 'Year', sector.id = 'All Occupations',
  start.year = 2013, end.year = 2015)
}
