% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_share.R
\name{shift_share}
\alias{shift_share}
\title{A Shift-Share Function}
\usage{
shift_share(ap_grp_labels, ap_numerator1, ap_numerator2, ap_denominator1,
  ap_denominator2)
}
\arguments{
\item{ap_grp_labels}{#groups}

\item{ap_numerator1}{#numerator of rate 1}

\item{ap_numerator2}{#numerator of rate 2}

\item{ap_denominator1}{#denominator of rate 1}

\item{ap_denominator2}{#denominator of rate 2}
}
\value{
object
shift_share()
}
\description{
This function allows you to do shift-share analysis on rates (e.g. employment rate, basketball field goal % for 2 point - 3 point analysis)
}
