% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{shiftshare_environment}
\alias{shiftshare_environment}
\title{Simulator of the Shift-Share Environment}
\usage{
shiftshare_environment(
  seed = 1,
  II = 100,
  NN = 30,
  effect_of_interest = 1,
  weight_on_IV = 0.5,
  eps_sd = 0.1,
  regionshock_sd = 0.1
)
}
\arguments{
\item{seed}{Sets the random seed.}

\item{II}{Number of locations to simulate.}
}
\description{
Simulator of the Shift-Share Environment
}
