% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{ssaggregate_check}
\alias{ssaggregate_check}
\title{Demonstrate that ssaggregate and our approach are identical.}
\usage{
ssaggregate_check(seed = 1)
}
\arguments{
\item{seed}{Random seed.}
}
\description{
Demonstrate that ssaggregate and our approach are identical.
}
